refactor history component to its own package
