+++ collated/main/A0142130A.md
@@ -394,9 +394,9 @@ import java.util.logging.Logger;
 import seedu.taskell.commons.core.EventsCenter;
 import seedu.taskell.commons.core.LogsCenter;
 import seedu.taskell.commons.events.ui.DisplayListChangedEvent;
-import seedu.taskell.model.CommandHistory;
-import seedu.taskell.model.History;
-import seedu.taskell.model.HistoryManager;
+import seedu.taskell.history.CommandHistory;
+import seedu.taskell.history.History;
+import seedu.taskell.history.HistoryManager;
 import seedu.taskell.model.Model;
 import seedu.taskell.model.task.Task;
 import seedu.taskell.model.task.UniqueTaskList.DuplicateTaskException;
@@ -604,8 +604,8 @@ package seedu.taskell.logic.commands;
 
 import seedu.taskell.commons.core.EventsCenter;
 import seedu.taskell.commons.events.ui.DisplayListChangedEvent;
-import seedu.taskell.model.History;
-import seedu.taskell.model.HistoryManager;
+import seedu.taskell.history.History;
+import seedu.taskell.history.HistoryManager;
 
 /** Lists a list of previous commands available for Undo operation
  * */
