+++ collated/test/A0142130A.md
@@ -154,8 +154,8 @@ import org.junit.Test;
 
 import guitests.guihandles.TaskCardHandle;
 import seedu.taskell.logic.commands.UndoCommand;
-import seedu.taskell.model.History;
-import seedu.taskell.model.HistoryManager;
+import seedu.taskell.history.History;
+import seedu.taskell.history.HistoryManager;
 import seedu.taskell.testutil.TestTask;
 import seedu.taskell.testutil.TestUtil;
 
@@ -302,7 +302,7 @@ import java.util.ArrayList;
 import org.junit.Test;
 
 import seedu.taskell.logic.commands.ViewHistoryCommand;
-import seedu.taskell.model.HistoryManager;
+import seedu.taskell.history.HistoryManager;
 import seedu.taskell.ui.DisplayPanel;
 
 public class ViewHistoryCommandTest extends TaskManagerGuiTest {
