+++ src/main/java/seedu/taskell/logic/commands/AddCommand.java
@@ -4,7 +4,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 import seedu.taskell.commons.exceptions.IllegalValueException;
-import seedu.taskell.model.HistoryManager;
+import seedu.taskell.history.HistoryManager;
 import seedu.taskell.model.tag.Tag;
 import seedu.taskell.model.tag.UniqueTagList;
 import seedu.taskell.model.task.*;
