+++ src/main/java/seedu/taskell/logic/commands/UndoCommand.java
@@ -7,9 +7,9 @@ import java.util.logging.Logger;
 import seedu.taskell.commons.core.EventsCenter;
 import seedu.taskell.commons.core.LogsCenter;
 import seedu.taskell.commons.events.ui.DisplayListChangedEvent;
-import seedu.taskell.model.CommandHistory;
-import seedu.taskell.model.History;
-import seedu.taskell.model.HistoryManager;
+import seedu.taskell.history.CommandHistory;
+import seedu.taskell.history.History;
+import seedu.taskell.history.HistoryManager;
 import seedu.taskell.model.Model;
 import seedu.taskell.model.task.Task;
 import seedu.taskell.model.task.UniqueTaskList.DuplicateTaskException;
