+++ src/main/java/seedu/taskell/logic/commands/ViewHistoryCommand.java
@@ -3,8 +3,8 @@ package seedu.taskell.logic.commands;
 
 import seedu.taskell.commons.core.EventsCenter;
 import seedu.taskell.commons.events.ui.DisplayListChangedEvent;
-import seedu.taskell.model.History;
-import seedu.taskell.model.HistoryManager;
+import seedu.taskell.history.History;
+import seedu.taskell.history.HistoryManager;
 
 /** Lists a list of previous commands available for Undo operation
  * */
